backend/requirements.txt
amqp==5.1.1
asgiref==3.6.0
asttokens==2.2.1
async-timeout==4.0.2
attrs==22.2.0
autobahn==23.1.2
Automat==22.10.0
backcall==0.2.0
billiard==3.6.4.0
celery==5.2.7
cffi==1.15.1
channels==4.0.0
channels-redis==4.1.0
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
constantly==15.1.0
cron-descriptor==1.2.35
cryptography==40.0.1
daphne==4.0.0
debugpy==1.6.6
decorator==5.1.1
Django==4.1.7
django-celery-beat==2.5.0
django-celery-results==2.5.0
django-cors-headers==3.14.0
django-extensions==3.2.1
django-timezone-field==5.0
django_debug_toolbar==3.8.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
dnspython==2.3.0
eventlet==0.33.3
executing==1.2.0
greenlet==2.0.2
hyperlink==21.0.0
idna==3.4
incremental==22.10.0
ipython==8.12.0
jedi==0.18.2
kombu==5.2.4
matplotlib-inline==0.1.6
msgpack==1.0.5
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.4.0
prompt-toolkit==3.0.38
psycopg2-binary==2.9.5
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
Pygments==2.14.0
PyJWT==2.6.0
pyOpenSSL==23.1.1
python-crontab==2.7.1
python-dateutil==2.8.2
pytz==2023.2
redis==4.5.4
service-identity==21.1.0
six==1.16.0
sqlparse==0.4.3
stack-data==0.6.2
traitlets==5.9.0
Twisted==22.10.0
txaio==23.1.1
typing_extensions==4.5.0
tzdata==2023.3
vine==5.0.0
wcwidth==0.2.6
zope.interface==6.0
